+++ packages/orq-rc/src/orq_ai_sdk/models/updatepromptsnippetop.py
@@ -1891,7 +1891,7 @@ class UpdatePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-11T12:04:07.091Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-11T13:33:05.305Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")